Refactor ReportControllerTest, PetOwnerControllerTest
