+++ src/test/java/pro/sky/animal_shelter_telegram_bot/controller/ConstantsForControllerTests.java
@@ -13,7 +13,6 @@ public class ConstantsForControllerTests {
     public final static String VOLUNTEER_URL = "volunteer";
     public final static String ALL_URL = "all";
 
-
     public final static Pet PET = new Pet();
     public final static PetOwner PET_OWNER = new PetOwner();
 
@@ -29,15 +28,4 @@ public class ConstantsForControllerTests {
     public final static String DATE_OF_REPORT = "12.12.2021";
     public final static String DIET = "Meat";
     public final static String RESULT = "Report is fine";
-
-    public final static String HELLO_MESSAGE_PET_OWNER = "You can do it by information of pet owner:\n" +
-            "1. add information about the owner of the pet\n" +
-            "2. get information about the owner of the pet\n" +
-            "2. update information about the owner of the pet\n" +
-            "4. remove information about the owner of the pet\n";
-    public final static String HELLO_MESSAGE_REPORT_CONTROLLER = "You can do it by reports\n" +
-            "1. add new report\n" +
-            "2. find report\n" +
-            "2. update report\n" +
-            "4. remove report\n";
 }
